\name{kinship-internal}
\alias{all}
\alias{any}
\alias{diag}
\alias{diag<-}
\alias{max}
\alias{min}
\alias{prod}
\alias{range}
\alias{sum}
\alias{unique}
\title{internal functions for kinship}
\description{internal kinship functions}
\details{These functions are not explicitly called by users}
\keyword{internal}
