\name{kinship.ops}
\alias{\%*\%}
\alias{\%*\%.gchol.bdsmatrix}
\alias{\%*\%.default}
\alias{\%*\%.bdsmatrix}
\alias{\%*\%-methods}
\alias{\%*\%,ANY,ANY-method}
\alias{\%*\%,bdsmatrix-method}
\alias{\%*\%,bdsmatrix,numeric-method}
\alias{\%*\%,gchol.bdsmatrix-method}
\alias{\%*\%,gchol.bdsmatrix,numeric-method}
\alias{[.pedigree}
\alias{as.matrix.gchol}
\alias{as.matrix.gchol.bdsmatrix}
\alias{as.vector.bdsmatrix}
\alias{bdsmatrix.reconcile}
\alias{is.list.bdsmatrix}
\alias{is.matrix.bdsmatrix}
\alias{is.matrix.gchol.bdsmatrix}
\alias{is.bdsmatrix}
\alias{is.gchol.bdsmatrix}
\alias{is.list.gchol.bdsmatrix}
\title{Basic Linear Algebra for classes bdsmatrix and gchol.bdsmatrix}
\description{Basic linear algebra operators for class \code{bdsmatrix} and 
\code{gchol.bdsmatrix}}
\keyword{algebra}
