\name{list or NULL-class}
\docType{class}
\alias{list or NULL-class}

\title{Class "list or NULL" }
\description{ a class used in bdsmatrix}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "list or NULL" in the signature.
}

\keyword{classes}
