\name{solve.gchol.bdsmatrix}
\alias{solve.gchol.bdsmatrix}
\title{ function solve for gchol.bdsmatrix }
\description{
  This is a generalized solve function for gchol.bdsmatrix object
}
\usage{
\method{solve}{gchol.bdsmatrix}(a, b, full=T, \dots)
}
\arguments{
  \item{a}{ \code{a} }
  \item{b}{ \code{b} }
  \item{full}{ An argument \code{full}}
  \item{\ldots}{an argument to achieve compatibility with solve from R base}
}

\keyword{algebra}
